Move AbsenceSettings into the absence package
