+++ src/main/java/org/synyx/urlaubsverwaltung/absence/AbsenceSettings.java
@@ -1,4 +1,4 @@
-package org.synyx.urlaubsverwaltung.settings;
+package org.synyx.urlaubsverwaltung.absence;
 
 import javax.persistence.Embeddable;
