+++ src/main/java/org/synyx/urlaubsverwaltung/sicknote/SickNoteServiceImpl.java
@@ -2,8 +2,8 @@ package org.synyx.urlaubsverwaltung.sicknote;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.synyx.urlaubsverwaltung.absence.AbsenceSettings;
 import org.synyx.urlaubsverwaltung.person.Person;
-import org.synyx.urlaubsverwaltung.settings.AbsenceSettings;
 import org.synyx.urlaubsverwaltung.settings.Settings;
 import org.synyx.urlaubsverwaltung.settings.SettingsService;
